+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 This file is used to list changes made in each version of grafana.
 
-## dev:
+## 2.1.1 (2015-08-20):
 
+* Update Grafana default version to 2.1.2
 * Ensure we're displaying better messages in some edge cases with HTTP requests [#76](https://github.com/JonathanTron/chef-grafana/issues/76)
 
 ## 2.1.0 (2015-08-17):